+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Teknoo Software - Kubernetes Client - Change Log
 
-## [0.30.0] - 2022-12-19
+## [0.30.1] - 2023-01-04
+### Dev Release
+- Client::setOptions (and Client's construction) require a master value to be executed
+- All 40x Errors returned by Kubernetes API throw now an exception  
+
+## [0.30.0] - 2022-12-31
 ### Dev Release
 - Fork from [Maclof Kuebrnetes library](https://github.com/maclof/kubernetes-client)
 - Rewrite to PSR-1, PSR-12, PSR-4.